+++ api_app/analyzers_manager/file_analyzers/hfinger.py
@@ -16,7 +16,17 @@ class Hfinger(FileAnalyzer):
     fingerprint_report_mode: int = 2
 
     def run(self):
-        return hfinger_analyze(self.filepath, self.fingerprint_report_mode)
+        reports = dict()
+        reports["extraction"] = hfinger_analyze(
+            self.filepath, self.fingerprint_report_mode
+        )
+        fingerprints = set()
+        for report in reports["extraction"]:
+            fingerprint = report.get("fingerprint", "")
+            if fingerprint:
+                fingerprints.add(fingerprint)
+        reports["fingerprints_summary"] = list(fingerprints)
+        return reports
 
     @classmethod
     def update(cls) -> bool: