+++ api_app/analyzers_manager/file_analyzers/triage_scan.py
@@ -26,7 +26,7 @@ class TriageScanFile(FileAnalyzer, TriageMixin):
         logger.info(f"triage md5 {self.md5} sending sample for analysis")
         for _try in range(self.max_tries):
             logger.info(f"triage md5 {self.md5} polling for result try #{_try + 1}")
-            self.response = self.session.post(self.base_url + "samples", files=files)
+            self.response = self.session.post(self.url + "samples", files=files)
             if self.response.status_code == 200:
                 break
             time.sleep(self.poll_distance)