+++ frontend/tests/utils/observables.test.js
@@ -18,6 +18,12 @@ describe("test observables utilities functions", () => {
     expect(getObservableClassification("hello world")).toBe(
       ObservableClassifications.GENERIC,
     );
+    expect(getObservableClassification("+391234567890")).toBe(
+      ObservableClassifications.GENERIC,
+    );
+    expect(getObservableClassification("2024-05-10")).toBe(
+      ObservableClassifications.GENERIC,
+    );
     expect(getObservableClassification("google.]com")).toBe(
       ObservableClassifications.DOMAIN,
     );
@@ -105,4 +111,57 @@ describe("Observable validators tests", () => {
       observable: "http://test.com",
     });
   });
+
+  test.each([
+    "+391234567890",
+    "+391234567890 ",
+    "+391234567890;",
+    " +391234567890 ",
+  ])("test valid phone numbers (%s)", (valueToValidate) => {
+    expect(observableValidators(valueToValidate)).toBeNull();
+  });
+
+  test.each([
+    "123 4567890",
+    "123-456-7890",
+    "(123) 456-7890",
+    "123 456 7890",
+    "123.456.7890",
+    "+91 (123) 456-7890",
+    "+39 123 4567890",
+    "12345 67890",
+    "123-4567890",
+  ])("test invalid phone numbers (%s)", (valueToValidate) => {
+    /* these are valid phone numbers format in the real world, but it's too much complex to support all of them.
+    Also store the same type of data in different format add complexity.
+    If we want to support them store in the db the phone numbers always in the same format.
+    */
+    const validationResult = observableValidators(valueToValidate);
+    /* some of the elements match the ip regex, in case they don't match it return null:
+     use the ?. to access to the field for the null element and it will be undefined.
+    */
+    expect(["ip", undefined]).toContain(validationResult?.classification);
+  });
+
+  test.each([
+    "2024-05-10",
+    "2024-05-10 ",
+    "2024-05-10;",
+    " 2024-05-10 ",
+    "10/10/21",
+    "2024-05-10T12:30:40Z",
+    "2024-05-10T12:30:40",
+    "2024-05-10 12:30:40",
+  ])("test valid date (%s)", (valueToValidate) => {
+    expect(observableValidators(valueToValidate)).toBeNull();
+  });
+
+  test.each(["2024-05-40", "10/13/21", "2024-05-10 56:30:40"])(
+    "test valid date (%s)",
+    (valueToValidate) => {
+      const validationResult = observableValidators(valueToValidate);
+      // 2024-05-40 matches IP addess regex.
+      expect(["ip", undefined]).toContain(validationResult?.classification);
+    },
+  );
 });